--- a/benchmark/tpch/csv/write_lineitem_csv_no_order.benchmark
+++ b/benchmark/tpch/csv/write_lineitem_csv_no_order.benchmark
@@ -1,0 +1,18 @@
+# name: benchmark/tpch/csv/write_lineitem_csv_no_order.benchmark
+# description: Write the lineitem of TPC-H SF0.1 to a CSV file (without preserving insertion order)
+# group: [csv]
+
+name Write Lineitem CSV (Non-Order Preserving)
+group csv
+
+require tpch
+
+load
+CALL dbgen(sf=0.1);
+SET preserve_insertion_order=false;
+
+run
+COPY lineitem TO '${BENCHMARK_DIR}/lineitem.csv' (FORMAT CSV, DELIMITER '|', HEADER);
+
+result I
+600572