--- a/benchmark/tpch/parquet/write_lineitem_parquet.benchmark
+++ b/benchmark/tpch/parquet/write_lineitem_parquet.benchmark
@@ -1,0 +1,19 @@
+# name: benchmark/tpch/parquet/write_lineitem_parquet.benchmark
+# description: Write the lineitem of TPC-H SF1 to a Parquet file
+# group: [parquet]
+
+name Write Lineitem Parquet
+group parquet
+
+require parquet
+
+require tpch
+
+load
+CALL dbgen(sf=1);
+
+run
+COPY lineitem TO '${BENCHMARK_DIR}/lineitem.parquet';
+
+result I
+6001215