--- a/benchmark/tpch/parquet/write_lineitem_parquet_no_order.benchmark
+++ b/benchmark/tpch/parquet/write_lineitem_parquet_no_order.benchmark
@@ -1,0 +1,20 @@
+# name: benchmark/tpch/parquet/write_lineitem_parquet_no_order.benchmark
+# description: Write the lineitem of TPC-H SF1 to a Parquet file
+# group: [parquet]
+
+name Write Lineitem Parquet (Non-Order Preserving)
+group parquet
+
+require parquet
+
+require tpch
+
+load
+CALL dbgen(sf=1);
+SET preserve_insertion_order=false;
+
+run
+COPY lineitem TO '${BENCHMARK_DIR}/lineitem.parquet';
+
+result I
+6001215