--- a/test/sql/copy/parquet/tpch_parallel_copy.test_slow
+++ b/test/sql/copy/parquet/tpch_parallel_copy.test_slow
@@ -1,0 +1,44 @@
+# name: test/sql/copy/parquet/tpch_parallel_copy.test_slow
+# description: Test parallel TPC-H SF0.1
+# group: [parquet]
+
+require parquet
+
+require tpch
+
+statement ok
+CALL dbgen(sf=0.1, suffix='_original');
+
+statement ok
+CALL dbgen(sf=0);
+
+statement ok
+SET preserve_insertion_order=false
+
+foreach tpch_tbl orders customer lineitem nation part partsupp region supplier
+
+statement ok
+COPY ${tpch_tbl}_original TO '__TEST_DIR__/${tpch_tbl}.parquet';
+
+statement ok
+COPY ${tpch_tbl} FROM '__TEST_DIR__/${tpch_tbl}.parquet';
+
+endloop
+
+loop i 1 9
+
+query I
+PRAGMA tpch(${i})
+----
+<FILE>:extension/tpch/dbgen/answers/sf0.1/q0${i}.csv
+
+endloop
+
+loop i 10 23
+
+query I
+PRAGMA tpch(${i})
+----
+<FILE>:extension/tpch/dbgen/answers/sf0.1/q${i}.csv
+
+endloop