--- a/test/sql/copy/parquet/tpch_parallel_copy.test_slow
+++ b/test/sql/copy/parquet/tpch_parallel_copy.test_slow
@@ -7,7 +7,7 @@ require parquet
 require tpch
 
 statement ok
-CALL dbgen(sf=0.1, suffix='_original');
+CALL dbgen(sf=1, suffix='_original');
 
 statement ok
 CALL dbgen(sf=0);
@@ -30,7 +30,7 @@ loop i 1 9
 query I
 PRAGMA tpch(${i})
 ----
-<FILE>:extension/tpch/dbgen/answers/sf0.1/q0${i}.csv
+<FILE>:extension/tpch/dbgen/answers/sf1/q0${i}.csv
 
 endloop
 
@@ -39,6 +39,6 @@ loop i 10 23
 query I
 PRAGMA tpch(${i})
 ----
-<FILE>:extension/tpch/dbgen/answers/sf0.1/q${i}.csv
+<FILE>:extension/tpch/dbgen/answers/sf1/q${i}.csv
 
 endloop