This PR adds support for parallel writing to CSV/Parquet files as long as order-preservation is turned off:

```sql
SET preserve_insertion_order=false;
```

Below are some timings of writing the TPC-H SF1 lineitem table to CSV/Parquet on my laptop:

| Format  | Old  | New (8T) |
|---------|------|----------|
| CSV     | 2.6s | 0.38s    |
| Parquet | 7.5s | 1.3s     |